+++ lib/injection_container.dart
@@ -13,14 +13,18 @@ import 'features/consulta_fipe/data/models/modelo_model.dart';
 import 'features/consulta_fipe/data/models/valor_fipe_model.dart';
 import 'features/consulta_fipe/data/repositories/fipe_repository_impl.dart';
 import 'features/consulta_fipe/domain/repositories/fipe_repository.dart';
+import 'features/consulta_fipe/domain/usecases/check_for_updates_usecase.dart';
 import 'features/consulta_fipe/domain/usecases/get_anos_combustiveis_por_modelo_usecase.dart';
 import 'features/consulta_fipe/domain/usecases/get_anos_por_marca_usecase.dart';
+import 'features/consulta_fipe/domain/usecases/get_local_mes_referencia_usecase.dart';
 import 'features/consulta_fipe/domain/usecases/get_marcas_por_tipo_usecase.dart';
 import 'features/consulta_fipe/domain/usecases/get_modelos_por_marca_usecase.dart';
 import 'features/consulta_fipe/domain/usecases/get_valor_fipe_usecase.dart';
+import 'features/consulta_fipe/domain/usecases/sync_all_data_usecase.dart';
 import 'features/consulta_fipe/presentation/bloc/ano_combustivel_bloc.dart';
 import 'features/consulta_fipe/presentation/bloc/marca_bloc.dart';
 import 'features/consulta_fipe/presentation/bloc/modelo_bloc.dart';
+import 'features/consulta_fipe/presentation/bloc/sync_bloc.dart';
 import 'features/consulta_fipe/presentation/bloc/valor_fipe_bloc.dart';
 
 /// Service Locator global
@@ -68,6 +72,13 @@ Future<void> initDependencies() async {
     ),
   );
   sl.registerFactory(() => ValorFipeBloc(getValorFipe: sl()));
+  sl.registerFactory(
+    () => SyncBloc(
+      checkForUpdates: sl(),
+      syncAllData: sl(),
+      getLocalMesReferencia: sl(),
+    ),
+  );
 
   // ! UseCases - Lazy Singleton
   sl.registerLazySingleton(() => GetMarcasPorTipoUseCase(sl()));
@@ -75,6 +86,9 @@ Future<void> initDependencies() async {
   sl.registerLazySingleton(() => GetAnosCombustiveisPorModeloUseCase(sl()));
   sl.registerLazySingleton(() => GetAnosPorMarcaUseCase(sl()));
   sl.registerLazySingleton(() => GetValorFipeUseCase(sl()));
+  sl.registerLazySingleton(() => CheckForUpdatesUseCase(sl()));
+  sl.registerLazySingleton(() => SyncAllDataUseCase(sl()));
+  sl.registerLazySingleton(() => GetLocalMesReferenciaUseCase(sl()));
 
   // ! Repositories - Lazy Singleton
   sl.registerLazySingleton<FipeRepository>(